Move to new statistics/ subfolder of storage
