+++ agent_api/src/main/java/dev/aikido/agent_api/background/cloud/api/events/Heartbeat.java
@@ -3,7 +3,7 @@ package dev.aikido.agent_api.background.cloud.api.events;
 import dev.aikido.agent_api.background.cloud.GetManagerInfo;
 import dev.aikido.agent_api.storage.Hostnames;
 import dev.aikido.agent_api.storage.ServiceConfigStore;
-import dev.aikido.agent_api.storage.Statistics;
+import dev.aikido.agent_api.storage.statistics.Statistics;
 import dev.aikido.agent_api.storage.routes.RouteEntry;
 import dev.aikido.agent_api.context.User;
